Casting string::npos to int in the if statement

+++ project/raytraceCuda.cpp
@@ -641,7 +641,7 @@ void getArguments(int argc, char* argv[])
         unsigned int isFile = firstArg.find(filetype);
         printf("location of '.txt' in firstArg: %d\n", isFile);
         printf("value of string::npos: %d\n", (int) string::npos);
-        if (isFile != string::npos)
+        if (isFile != (int) string::npos)
         {
             parseFile(argv[1]);
         }
